refactor: renames "index.ts" to "exports.ts" in "library/Attempt/Outcome"
