+++ src/library/Attempt/ended.ts
@@ -1,6 +1,6 @@
 import {
   Attempt_Outcome,
-} from './Outcome';
+} from './Outcome/exports';
 
 import {
   NonActionableError,
