+++ src/library/Attempt/exports.ts
@@ -12,7 +12,7 @@ export {
   Attempt_Outcome as Outcome,
   type Attempt_Success as Success,
   type Attempt_Failure as Failure,
-} from './Outcome';
+} from './Outcome/exports';
 
 export {
   attemptTo as to,
